# Conflicts:
#	SofaKernel/framework/sofa/core/behavior/BaseConstraintCorrection.h
#	SofaKernel/framework/sofa/core/behavior/LinearSolver.h
#	SofaKernel/framework/sofa/simulation/Node.cpp
#	SofaKernel/modules/SofaBaseLinearSolver/MatrixLinearSolver.h
#	modules/SofaConstraint/ConstraintStoreLambdaVisitor.h
#	modules/SofaConstraint/GenericConstraintCorrection.h
#	modules/SofaConstraint/LinearSolverConstraintCorrection.inl
#	modules/SofaConstraint/PrecomputedConstraintCorrection.h
#	modules/SofaConstraint/UncoupledConstraintCorrection.h
#	modules/SofaConstraint/UncoupledConstraintCorrection.inl

+++ modules/SofaSparseSolver/src/SofaSparseSolver/SparseLDLSolver.inl
@@ -1,6 +1,6 @@
 /******************************************************************************
-*       SOFA, Simulation Open-Framework Architecture, development version     *
-*                (c) 2006-2018 INRIA, USTL, UJF, CNRS, MGH                    *
+*                 SOFA, Simulation Open-Framework Architecture                *
+*                    (c) 2006 INRIA, USTL, UJF, CNRS, MGH                     *
 *                                                                             *
 * This program is free software; you can redistribute it and/or modify it     *
 * under the terms of the GNU Lesser General Public License as published by    *
@@ -19,8 +19,6 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-// Author: Hadrien Courtecuisse
-
 #ifndef SOFA_COMPONENT_LINEARSOLVER_SparseLDLSolver_INL
 #define SOFA_COMPONENT_LINEARSOLVER_SparseLDLSolver_INL
 
@@ -30,7 +28,7 @@
 #include "sofa/helper/system/thread/CTime.h"
 #include <sofa/core/objectmodel/BaseContext.h>
 #include <sofa/core/behavior/LinearSolver.h>
-#include <math.h>
+#include <cmath>
 #include <sofa/helper/system/thread/CTime.h>
 #include <SofaBaseLinearSolver/CompressedRowSparseMatrix.inl>
 #include <fstream>
@@ -82,7 +80,7 @@ void SparseLDLSolver<TMatrix,TVector,TThreadManager>::invert(Matrix& M) {
     int * M_rowind = (int *) &Mfiltered.getColsIndex()[0];
     Real * M_values = (Real *) &Mfiltered.getColsValue()[0];
 
-    if(M_colptr==nullptr || M_rowind==nullptr || M_values==nullptr || Mfiltered.getRowBegin().size() < n )
+    if(M_colptr==nullptr || M_rowind==nullptr || M_values==nullptr || Mfiltered.getRowBegin().size() < (size_t)n )
     {
         msg_warning() << "Invalid Linear System to solve. Please insure that there is enough constraints (not rank deficient)." ;
         return ;
